requirements.txt
asgiref==3.5.0
astroid==2.9.3
atomicwrites==1.4.0
attrs==21.4.0
black==22.1.0
cfgv==3.3.1
click==8.0.4
colorama==0.4.4
coverage==6.3.2
distlib==0.3.4
Django==4.0.2
django-environ==0.8.1
djangorestframework==3.13.1
filelock==3.6.0
identify==2.4.11
iniconfig==1.1.1
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.6.1
model-bakery==1.4.0
mypy-extensions==0.4.3
nodeenv==1.6.0
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.1
pluggy==1.0.0
pre-commit==2.17.0
py==1.11.0
pylint==2.12.2
pyparsing==3.0.7
pytest==7.0.1
pytest-django==4.5.2
pytz==2021.3
PyYAML==6.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
tzdata==2021.5
virtualenv==20.13.2
wrapt==1.13.3
